+++ docs/pydl/changes.rst
@@ -25,6 +25,8 @@ PyDL Changelog
 * Uppercase the package name (in documentation only).
 * Upgrade to `astropy_helpers`_/v1.1.1.
 * Add functions from the idlutils rgbcolor directory.
+* :func:`~pydl.pydlspec2d.spec1d.spec_path` can now find SDSS spectra, not just
+  BOSS.
 
 .. _`PEP 8`: https://www.python.org/dev/peps/pep-0008/
 .. _`PR #7`: https://github.com/weaverba137/pydl/pull/7